tags: prepare tag types, functions, and methods to be move to a tags pkg
Rename: tagPair => Tag; tagSet => TagSet

Make all tag functions stand-alone (aka not subScope methods).
