+++ stats.go
@@ -380,7 +380,7 @@ func (s *statStore) NewCounterWithTags(name string, tags map[string]string) Coun
 	return s.newCounter(serializeTags(name, tags))
 }
 
-func (s *statStore) newCounterWithTagSet(name string, tags tagSet) Counter {
+func (s *statStore) newCounterWithTagSet(name string, tags TagSet) Counter {
 	return s.newCounter(serializeTagSet(name, tags))
 }
 
@@ -392,12 +392,12 @@ var emptyPerInstanceTags = map[string]string{"_f": "i"}
 // Previously, we modified the map, but that is no longer allowed and creating
 // a new tagSet is faster and more memory efficient than creating a duplicate
 // map.
-func convertTagsToPerInstanceTagSet(tags map[string]string) tagSet {
-	set := make(tagSet, 1, len(tags)+1)
-	set[0] = tagPair{key: "_f", value: "i"}
+func convertTagsToPerInstanceTagSet(tags map[string]string) TagSet {
+	set := make(TagSet, 1, len(tags)+1)
+	set[0] = Tag{key: "_f", value: "i"}
 	for k, v := range tags {
 		if k != "" && v != "" {
-			set = append(set, tagPair{key: k, value: replaceChars(v)})
+			set = append(set, Tag{key: k, value: replaceChars(v)})
 		}
 	}
 	set.Sort()
@@ -433,7 +433,7 @@ func (s *statStore) NewGaugeWithTags(name string, tags map[string]string) Gauge 
 	return s.newGauge(serializeTags(name, tags))
 }
 
-func (s *statStore) newGaugeWithTagSet(name string, tags tagSet) Gauge {
+func (s *statStore) newGaugeWithTagSet(name string, tags TagSet) Gauge {
 	return s.newGauge(serializeTagSet(name, tags))
 }
 
@@ -466,7 +466,7 @@ func (s *statStore) NewTimerWithTags(name string, tags map[string]string) Timer 
 	return s.newTimer(serializeTags(name, tags))
 }
 
-func (s *statStore) newTimerWithTagSet(name string, tags tagSet) Timer {
+func (s *statStore) newTimerWithTagSet(name string, tags TagSet) Timer {
 	return s.newTimer(serializeTagSet(name, tags))
 }
 
@@ -483,14 +483,14 @@ func (s *statStore) NewPerInstanceTimer(name string, tags map[string]string) Tim
 type subScope struct {
 	registry *statStore
 	name     string
-	tags     tagSet // read-only and may be shared by multiple subScopes
+	tags     TagSet // read-only and may be shared by multiple subScopes
 }
 
 func newSubScope(registry *statStore, name string, tags map[string]string) *subScope {
-	a := make(tagSet, 0, len(tags))
+	a := make(TagSet, 0, len(tags))
 	for k, v := range tags {
 		if k != "" && v != "" {
-			a = append(a, tagPair{key: k, value: replaceChars(v)})
+			a = append(a, Tag{key: k, value: replaceChars(v)})
 		}
 	}
 	a.Sort()
@@ -505,7 +505,7 @@ func (s *subScope) ScopeWithTags(name string, tags map[string]string) Scope {
 	return &subScope{
 		registry: s.registry,
 		name:     joinScopes(s.name, name),
-		tags:     s.mergeTags(tags),
+		tags:     mergeTags(s.tags, tags),
 	}
 }
 
@@ -518,12 +518,12 @@ func (s *subScope) NewCounter(name string) Counter {
 }
 
 func (s *subScope) NewCounterWithTags(name string, tags map[string]string) Counter {
-	return s.registry.newCounterWithTagSet(joinScopes(s.name, name), s.mergeTags(tags))
+	return s.registry.newCounterWithTagSet(joinScopes(s.name, name), mergeTags(s.tags, tags))
 }
 
 func (s *subScope) NewPerInstanceCounter(name string, tags map[string]string) Counter {
 	return s.registry.newCounterWithTagSet(joinScopes(s.name, name),
-		s.mergePerInstanceTags(tags))
+		mergePerInstanceTags(s.tags, tags))
 }
 
 func (s *subScope) NewGauge(name string) Gauge {
@@ -531,12 +531,12 @@ func (s *subScope) NewGauge(name string) Gauge {
 }
 
 func (s *subScope) NewGaugeWithTags(name string, tags map[string]string) Gauge {
-	return s.registry.newGaugeWithTagSet(joinScopes(s.name, name), s.mergeTags(tags))
+	return s.registry.newGaugeWithTagSet(joinScopes(s.name, name), mergeTags(s.tags, tags))
 }
 
 func (s *subScope) NewPerInstanceGauge(name string, tags map[string]string) Gauge {
 	return s.registry.newGaugeWithTagSet(joinScopes(s.name, name),
-		s.mergePerInstanceTags(tags))
+		mergePerInstanceTags(s.tags, tags))
 }
 
 func (s *subScope) NewTimer(name string) Timer {
@@ -544,108 +544,14 @@ func (s *subScope) NewTimer(name string) Timer {
 }
 
 func (s *subScope) NewTimerWithTags(name string, tags map[string]string) Timer {
-	return s.registry.newTimerWithTagSet(joinScopes(s.name, name), s.mergeTags(tags))
+	return s.registry.newTimerWithTagSet(joinScopes(s.name, name), mergeTags(s.tags, tags))
 }
 
 func (s *subScope) NewPerInstanceTimer(name string, tags map[string]string) Timer {
 	return s.registry.newTimerWithTagSet(joinScopes(s.name, name),
-		s.mergePerInstanceTags(tags))
+		mergePerInstanceTags(s.tags, tags))
 }
 
 func joinScopes(parent, child string) string {
 	return parent + "." + child
-}
-
-// mergeOneTag is an optimized for inserting 1 tag and will panic otherwise.
-func (s *subScope) mergeOneTag(tags map[string]string) tagSet {
-	if len(tags) != 1 {
-		panic("invalid usage")
-	}
-	var p tagPair
-	for k, v := range tags {
-		p = tagPair{key: k, value: replaceChars(v)}
-		break
-	}
-	if p.key == "" || p.value == "" {
-		return s.tags
-	}
-	a := make(tagSet, len(s.tags), len(s.tags)+1)
-	copy(a, s.tags)
-	return a.Insert(p)
-}
-
-// mergeTags returns a tagSet that is the union of subScope's tags and the
-// provided tags map. If any keys overlap the values from the provided map
-// are used.
-func (s *subScope) mergeTags(tags map[string]string) tagSet {
-	switch len(tags) {
-	case 0:
-		return s.tags
-	case 1:
-		// optimize for the common case of there only being one tag
-		return s.mergeOneTag(tags)
-	default:
-		// write tags to the end of the scratch slice
-		scratch := make(tagSet, len(s.tags)+len(tags))
-		a := scratch[len(s.tags):]
-		i := 0
-		for k, v := range tags {
-			if k != "" && v != "" {
-				a[i] = tagPair{key: k, value: replaceChars(v)}
-				i++
-			}
-		}
-		a = a[:i]
-		a.Sort()
-
-		if len(s.tags) == 0 {
-			return a
-		}
-		return mergeTagSets(s.tags, a, scratch)
-	}
-}
-
-// mergePerInstanceTags returns a tagSet that is the union of subScope's
-// tags and the provided tags map with. If any keys overlap the values from
-// the provided map are used.
-//
-// The returned tagSet will have a per-instance key ("_f") and if neither the
-// subScope or tags have this key it's value will be the default per-instance
-// value ("i").
-//
-// The method does not optimize for the case where there is only one tag
-// because it is used less frequently.
-func (s *subScope) mergePerInstanceTags(tags map[string]string) tagSet {
-	if len(tags) == 0 {
-		if s.tags.Contains("_f") {
-			return s.tags
-		}
-		// create copy with the per-instance tag
-		a := make(tagSet, len(s.tags), len(s.tags)+1)
-		copy(a, s.tags)
-		return a.Insert(tagPair{key: "_f", value: "i"})
-	}
-
-	// write tags to the end of scratch slice
-	scratch := make(tagSet, len(s.tags)+len(tags)+1)
-	a := scratch[len(s.tags):]
-	i := 0
-	for k, v := range tags {
-		if k != "" && v != "" {
-			a[i] = tagPair{key: k, value: replaceChars(v)}
-			i++
-		}
-	}
-	// add the default per-instance tag if not present
-	if tags["_f"] == "" && !s.tags.Contains("_f") {
-		a[i] = tagPair{key: "_f", value: "i"}
-		i++
-	}
-	a = a[:i]
-	a.Sort()
-
-	if len(s.tags) == 0 {
-		return a
-	}
-	return mergeTagSets(s.tags, a, scratch)
 }
